Add files via upload
This project is divided into 6 parts. In section 2, we download and extract a zip file " Penn-Fudan Database pedestrian detection and segmentation" which contains 170 images with 345 instances of pedestrians. In this project, we have only one class(a class for detecting persons) and 4 coordinates for drawing the bounding box. In this case, we should consider an annotation file, then write a custom dataset for it, and then visualize images with ground truth and semantic segmentation. After that, we prepare data using helper function and transform data then specify and split training and test set. Then download a pre-trained model (Faster R-CNN model with a Resnet-50 backbone) and adjust the input layer to resize the images into 224X224 pixels resolution and fine-tune the pre-trained model to our classes which is two (one class is for the person and the other is for background). In sections 4 and 5, after defining hyperparameters, we execute training and demonstrate multi-task losses in the plots and finally, we visualize the predicted bounding boxes and compare it with ground truth.